Crate pallet_detach

This pallet detach performs logic related to Detaching a Proto-Fragment from the Clamor Blockchain to an External Blockchain

Re-exports

pub use pallet::*;

Modules

crypto

Based on the above KeyTypeId we need to generate a pallet-specific crypto type wrappers. We can use from supported crypto kinds (sr25519, ed25519 and ecdsa) and augment the types with this pallet-specific identifier.

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

DetachInternalData

Payload represents information about a “detachable thing” (e.g a Proto-Fragment or a Fragment Instance) that will be detached

DetachRequest

Struct that represents a request to detach a “detachable thing” (e.g a Proto-Fragment or a Fragment Instance) from the Clamor Blockchain

ExportData

Struct that contains information about a “detached thing” (e.g a detached Proto-Fragment or a detached Fragment Instance) that was detached from the Clamor Blockchain

Enums

DetachHash

Enum representing a Proto-Fragment or a Fragment Instance that the User wants to detach from the Clamor Blockchain

SupportedChains

Possible Blockchains into which a Proto-Fragment can be detached

Constants

KEY_TYPE

Defines application identifier for crypto keys of this module.

Traits

WeightInfo

Weight functions needed for pallet_detach.