Module pallet_detach::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Re-exports

pub use __tt_error_token_1 as tt_error_token;

pub use __tt_default_parts_7 as tt_default_parts;

Structs

GenesisConfig

The Genesis Configuration for the Pallet.

Pallet

The pallet implementing the on-chain logic.

Enums

Call

Functions that are callable from outside the runtime.

Error

Custom dispatch errors of this pallet.

Event

The event emitted by this pallet.

Traits

Config

Configure the pallet by specifying the parameters and types on which it depends.

Type Definitions

DetachNonces

StorageDoubleMap that maps an account address on an external blockchain and the external blockchain itself to a nonce

DetachRequests

StorageValue that equals a list of detach requests

DetachedHashes

StorageMap that maps a detached Proto-Fragment’s hash to an ExportData enum (this enum contains information about the Proto-Fragment’ detachment)

EthereumAuthorities

StorageValue that equals the set of ECDSA public keys of the Ethereum accounts that are authorized to detach a Proto-Fragment onto Fragnova’s Ethereum Smart Contract

FragKeys

StorageValue that equals the List of Clamor Account IDs that both validate and send unsigned transactions with signed payload NOTE: Only the Root User of the Clamor Blockchain (i.e the local node itself) can edit FragKeys

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.