Crate pallet_detach

This pallet detach performs logic related to Detaching a Proto-Fragment from the Clamor Blockchain to an External Blockchain

Re-exports

pub use pallet::*;

Modules

crypto

Add a crypto module with an ed25519 signature key to ensure that your pallet owns an account that can be used for signing transactions.

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

DetachInternalData

Struct that represents a signed request (signed by the Clamor Account Address public) to detach a Proto-Fragment

DetachRequest

Struct that represents a request to detach a Proto-Fragment

ExportData

Struct that contains information about a detached Proto-Fragment

Enums

SupportedChains

**Possible Blockchains into which a Proto-Fragment can be detached

Constants

KEY_TYPE

Defines application identifier for crypto keys of this module.

Traits

WeightInfo

Weight functions needed for pallet_detach.